+++ src/Commands/Deployment/Log.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Sven\ForgeCLI\Commands\Deployment;
+
+use Sven\ForgeCLI\Commands\BaseCommand;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class Log extends BaseCommand
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function configure()
+    {
+        $this->setName('site:deploy-log')
+            ->addArgument('server', InputArgument::REQUIRED, 'The id of the server the site is on.')
+            ->addArgument('site', InputArgument::REQUIRED, 'The id of the site to get the deployment log for.')
+            ->setDescription('Get latest deployment log of the given site.');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function perform(InputInterface $input, OutputInterface $output)
+    {
+        $this->forge->siteDeploymentLog(
+            $input->getArgument('server'), $input->getArgument('site')
+        );
+    }
+}