+++ src/Commands/Deployment/ResetStatus.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Sven\ForgeCLI\Commands\Deployment;
+
+use Sven\ForgeCLI\Commands\BaseCommand;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class ResetStatus extends BaseCommand
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function configure()
+    {
+        $this->setName('site:reset-deployment-state')
+            ->addArgument('server', InputArgument::REQUIRED, 'The id of the server the site is on.')
+            ->addArgument('site', InputArgument::REQUIRED, 'The id of the site you want to reset the deployment state of.')
+            ->setDescription('Reset the deployment state of the given site.');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function perform(InputInterface $input, OutputInterface $output)
+    {
+        $this->forge->resetDeploymentState(
+            $input->getArgument('server'), $input->getArgument('site')
+        );
+    }
+}